requirements.txt
# CERVO BOT - Dependencias
# Flask y servidor
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1

# Requests para API calls
requests==2.31.0

# Google Gemini AI
google-generativeai==0.3.2

# Variables de entorno
python-dotenv==1.0.0

# Utilidades
urllib3==2.1.0
